+++ google2/CHANGELOG.md
@@ -9,6 +9,7 @@ Changed:
 - Add `getDisk`
 - Make `genDiskName` non-empty
 - Bump `http4s` version to `0.21.5`, `scalatest` to `3.1.2`, `fs2-io` to `2.4.2`
+- Add `autoDeleteDiskDeviceName: Set[DeviceName]` to `deleteInstance` method
 
 Added:
 - Add `GoogleDiskService` and `GoogleDiskInterpreter`
@@ -18,7 +19,7 @@ Added:
 - Add Generator for `DiskName`
 - Add Kubernetes client APIs for creating service accounts, roles and role bindings
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.10-3576093"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.10-TRAVIS-REPLACE-ME"`
 
 ## 0.9
 Changed: